SQLite Capability to Store Alert Requests
Add code to the app to implement SQLAlchemy in creating a table if it
does not exist, and then inserting in to a table that tracks the
alerts created by the user. Modifies the docker file and docker compose
file so that a new volume is created for the app to store the SQLite
database.

JIRA-7001

Add code to the app to implement SQLAlchemy in creating a table if it
does not exist, and then inserting in to a table that tracks the
alerts created by the user. Modifies the docker file and docker compose
file so that a new volume is created for the app to store the SQLite
database.